requirements.txt
Django>=5.0.2
djangorestframework>=3.14.0
Pillow>=10.2.0
django-cors-headers>=4.3.1
django-filter>=23.5
django-ckeditor>=6.7.0
django-meta>=2.3.1
django-taggit>=5.0.1
python-dotenv>=1.0.1
django-cleanup>=8.0.0
django-allauth>=0.61.1
django-redis>=5.4.0
